feat: useEffect를 통해서 주기적으로 렌더링을 하던 로직을 커스텀 훅으로 분리
